Fix cluster definition errors
* `min_workers` and `max_workers` should be int32, not int64
* `init_scripts` should be an array of `InitScriptInfo`

+++ models/auto_scale.go
@@ -15,10 +15,10 @@ import (
 type AutoScale struct {
 
 	// max workers
-	MaxWorkers int64 `json:"max_workers,omitempty"`
+	MaxWorkers int32 `json:"max_workers,omitempty"`
 
 	// min workers
-	MinWorkers int64 `json:"min_workers,omitempty"`
+	MinWorkers int32 `json:"min_workers,omitempty"`
 }
 
 // Validate validates this auto scale
